# ent/schema/user.go
package schema

import (
	"github.com/facebook/ent"
	"github.com/facebook/ent/schema/field"
)

// User holds the schema definition for the User entity.
type User struct {
	ent.Schema
}

// Fields of the User.
func (User) Fields() []ent.Field {
	return []ent.Field{
		field.Int("age").Positive(),
		field.String("name"),
		field.Bool("is_deleted").Default(false),
	}
}

// Edges of the User.
func (User) Edges() []ent.Edge {
	return nil
}